requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic
python-multipart # For FileResponse and form data if needed later
jinja2 # For nice error templates or HTML responses, optional but good for default exc handlers

# Testing dependencies
pytest
pytest-cov
pytest-env # To manage environment variables for tests (e.g., test DB config)
python-dotenv # For loading .env files if used for test DB config
passlib[bcrypt] # For password hashing
starlette # For SessionMiddleware (FastAPI is built on Starlette)
itsdangerous # For signing session cookies (often a dependency of SessionMiddleware)
```
